Terceiro commit

+++ src/pages/modal-materia/modal-materia.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { IonicPage, ViewController, ModalController, NavParams, ActionSheetController } from 'ionic-angular';
+import { IonicPage, ViewController, ModalController, NavParams, ActionSheetController, ToastController } from 'ionic-angular';
 import { DataBaseProvider } from '../../providers/database/database';
 import { Md5 } from 'ts-md5/dist/md5';
 import { File } from '@ionic-native/file';
@@ -18,6 +18,8 @@ export class ModalMateriaPage {
   private controleEditar;
   public materiaEditar;
   public listaHorario = [];
+  private nomeAntesEditar;
+
 
   public materia = {
     id: "",
@@ -30,21 +32,19 @@ export class ModalMateriaPage {
         id: "",
         inicio: "",
         fim: "",
-        dia: [{
-          dom: false,
-          seg: false,
-          ter: false,
-          qua: false,
-          qui: false,
-          sex: false,
-          sab: false
-        }],
+        dom: false,
+        seg: false,
+        ter: false,
+        qua: false,
+        qui: false,
+        sex: false,
+        sab: false,
         diasFormatados: ""
       }
     ]
   }
 
-  constructor(private view: ViewController, public actionsheetCtrl: ActionSheetController, private modal: ModalController, private dataBase: DataBaseProvider, params: NavParams, private file: File) {
+  constructor(private view: ViewController, public toastCtrl: ToastController, public actionsheetCtrl: ActionSheetController, private modal: ModalController, private dataBase: DataBaseProvider, params: NavParams, private file: File) {
 
     this.materiaEditar = params.get('materia');
 
@@ -56,6 +56,7 @@ export class ModalMateriaPage {
       this.materia = this.materiaEditar;
       this.listaHorario = this.materia.horario;
       this.controleEditar = true;
+      this.nomeAntesEditar = this.materiaEditar.nome;
     }
   }
 
@@ -66,9 +67,34 @@ export class ModalMateriaPage {
   async salvarHorario(metodo?: string) {
 
     try {
+      if (this.materia.nome === undefined || this.materia.nome === "" || this.materia.nome.trim() === "") {
+        this.presentToast('Nome é obrigatório!');
+        return;
+      }
+      this.materia.nome = this.materia.nome.trim();
+      let registroJaExiste = await this.isRegistroDuplicado(this.materia.nome);
+
+      let retornoBDHorarios = this.dataBase.getMaterias();
+      let horariosValidos = this.validaHorarioCadastradoEmOutraMateria(retornoBDHorarios, this.listaHorario);
+      if (!horariosValidos) {
+        return;
+      }
 
       if (this.controleEditar === true) {
+        // alert("Nome Antes: " + this.nomeAntesEditar + " Nome Depois: " + this.materiaEditar.nome)
+        if (this.materiaEditar.nome !== this.nomeAntesEditar) {
+          if (registroJaExiste) {
+            this.presentToast('Ja existe uma materia cadastrada com esse nome!');
+            return;
+          }
+        }
         await this.dataBase.removeMateria("id", this.materiaEditar.id);
+      }
+      else {
+        if (registroJaExiste) {
+          this.presentToast('Ja existe uma materia cadastrada com esse nome!');
+          return;
+        }
       }
 
       let retornoBD = this.dataBase.getMaterias();
@@ -83,9 +109,14 @@ export class ModalMateriaPage {
 
       if (this.controleEditar === false) {
         this.materia.id = Md5.hashStr((JSON.stringify(this.materia)) + new Date().toISOString()).toString();
+        // Verifica se pasta ja existe para salvar as fotos, se nao cria.
+        this.verificaDiretorio(this.materia.nome);
+      } else {
+        //alert("NomeAntes: " + this.nomeAntesEditar + "NomeNovo: " +this.materiaEditar.nome)
+        if (!this.materiaEditar.nome === this.nomeAntesEditar) {
+          this.renomeiaDiretorioMateria(this.nomeAntesEditar, this.materia.nome);
+        }
       }
-      // Verifica se pasta ja existe para salvar as fotos, senao cria.
-      this.verificaDiretorio(this.materia.nome);
 
       retornoBD.materias.push(this.materia);
 
@@ -100,6 +131,70 @@ export class ModalMateriaPage {
     }
   }
 
+  validaHorarioCadastradoEmOutraMateria(retBD, listHoras) {
+    if (retBD.materias !== undefined) {
+
+      const diasSemana = ["dom", "seg", "ter", "qua", "qui", "sex", "sab"];
+      let dataSalvandoInico = new Date();
+      let dataSalvandoFim = new Date();
+      let dataJaCadastradaInicio = new Date()
+      let dataJaCadastradaFim = new Date()
+
+      for (let materiaBD of retBD.materias) {
+        if (materiaBD.horario !== undefined) {
+          for (let horarioBD of materiaBD.horario) {
+            for (let materiaNova of listHoras) {
+              for (let dia of diasSemana) {
+                //alert("materiaBD[" + dia + "]: " + horarioBD[dia] + " materiaNova[" + dia + "]" + materiaNova[dia])
+                if (horarioBD[dia] && materiaNova[dia]) {
+
+                  //if (horarioBD[dia] && horarioSendoSalvo[dia]) {
+                  dataSalvandoInico.setHours(materiaNova["inicio"].split(":", 2)[0])
+                  dataSalvandoInico.setMinutes(materiaNova["inicio"].split(":", 2)[1])
+                  dataSalvandoInico.setSeconds(0);
+                  dataSalvandoFim.setHours(materiaNova["fim"].split(":", 2)[0])
+                  dataSalvandoFim.setMinutes(materiaNova["fim"].split(":", 2)[1])
+                  dataSalvandoFim.setSeconds(0);
+
+                  dataJaCadastradaInicio.setHours(horarioBD["inicio"].split(":", 2)[0])
+                  dataJaCadastradaInicio.setMinutes(horarioBD["inicio"].split(":", 2)[1])
+                  dataJaCadastradaInicio.setSeconds(0);
+                  dataJaCadastradaFim.setHours(horarioBD["fim"].split(":", 2)[0])
+                  dataJaCadastradaFim.setMinutes(horarioBD["fim"].split(":", 2)[1])
+                  dataJaCadastradaFim.setSeconds(0);
+
+                  if ((dataSalvandoInico > dataJaCadastradaInicio) && (dataSalvandoInico < dataJaCadastradaFim)) {
+                    this.presentToast('Horário em conflito com outro já cadastrado na matéria ' + JSON.stringify(materiaBD.nome));
+                    return false;
+                  }
+                  else if ((dataSalvandoFim > dataJaCadastradaInicio) && (dataSalvandoFim < dataJaCadastradaFim)) {
+                    this.presentToast('Horário em conflito com outro já cadastrado na matéria ' + JSON.stringify(materiaBD.nome));
+                    return false;
+                  }
+                  else if ((dataSalvandoInico < dataJaCadastradaInicio) && (dataSalvandoFim > dataJaCadastradaFim)) {
+                    this.presentToast('Horário em conflito com outro já cadastrado na matéria ' + JSON.stringify(materiaBD.nome));
+                    return false;
+                  }
+                  else if ((dataSalvandoInico > dataJaCadastradaInicio) && (dataSalvandoFim < dataJaCadastradaFim)) {
+                    this.presentToast('Horário em conflito com outro já cadastrado na matéria ' + JSON.stringify(materiaBD.nome));
+                    return false;
+                  }
+                  // Nao pode mesmo dia da semana com horario igual
+                  else if ((dataSalvandoInico.toString() === dataJaCadastradaInicio.toString()) || (dataSalvandoFim.toString() === dataJaCadastradaFim.toString())) {
+                    this.presentToast('Horário em conflito com outro já cadastrado na matéria ' + JSON.stringify(materiaBD.nome));
+                    return false;
+                  }
+
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+    return true;
+  }
+
   abrirNovoHorario() {
     const modal = this.modal.create('ModalMateriaHorarioPage');
 
@@ -111,23 +206,97 @@ export class ModalMateriaPage {
 
         let virgula = false;
 
-        if (data.dia.dom) { novoHorario = "Dom"; virgula = true }
-        if (data.dia.seg) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Seg"; virgula = true }
-        if (data.dia.ter) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Ter"; virgula = true }
-        if (data.dia.qua) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Qua"; virgula = true }
-        if (data.dia.qui) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Qui"; virgula = true }
-        if (data.dia.sex) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Sex"; virgula = true }
-        if (data.dia.sab) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Sab"; virgula = true }
+        if (data.dom) { novoHorario = "Dom"; virgula = true }
+        if (data.seg) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Seg"; virgula = true }
+        if (data.ter) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Ter"; virgula = true }
+        if (data.qua) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Qua"; virgula = true }
+        if (data.qui) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Qui"; virgula = true }
+        if (data.sex) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Sex"; virgula = true }
+        if (data.sab) { if (virgula) { novoHorario = novoHorario + ", " } novoHorario = novoHorario + "Sab"; virgula = true }
 
         data.diasFormatados = novoHorario;
 
         this.materia.horario = (data);
-        this.listaHorario.push(this.materia.horario)
+
+        let horarioValido = this.isHorarioValido(this.listaHorario, this.materia.horario)
+        if (horarioValido) {
+          this.listaHorario.push(this.materia.horario)
+        }
+
       }
 
     })
 
     modal.present();
+  }
+
+  isHorarioValido(harariosJainclusos, horarioSendoSalvo) {
+
+    // let dataSalvandoInico1 = new Date();
+    // let dataSalvandoFim1 = new Date();
+
+    // dataSalvandoInico1.setHours(horarioSendoSalvo["inicio"].split(":", 2)[0])
+    // dataSalvandoInico1.setMinutes(horarioSendoSalvo["inicio"].split(":", 2)[1])
+    // dataSalvandoInico1.setSeconds(0);
+    // dataSalvandoFim1.setHours(horarioSendoSalvo["fim"].split(":", 2)[0])
+    // dataSalvandoFim1.setMinutes(horarioSendoSalvo["fim"].split(":", 2)[1])
+    // dataSalvandoFim1.setSeconds(0);
+
+    // if (dataSalvandoFim1 < dataSalvandoInico1) {
+    //   this.presentToast('Data Final não pode ser menor que a Data Inicial!');
+    //   return false;
+    // }
+
+    const diasSemana = ["dom", "seg", "ter", "qua", "qui", "sex", "sab"];
+
+    for (let horarioBD of harariosJainclusos) {
+
+      let dataSalvandoInico = new Date();
+      let dataSalvandoFim = new Date();
+      let dataJaCadastradaInicio = new Date()
+      let dataJaCadastradaFim = new Date()
+
+      for (let dia of diasSemana) {
+        if (horarioBD[dia] && horarioSendoSalvo[dia]) {
+          dataSalvandoInico.setHours(horarioSendoSalvo["inicio"].split(":", 2)[0])
+          dataSalvandoInico.setMinutes(horarioSendoSalvo["inicio"].split(":", 2)[1])
+          dataSalvandoInico.setSeconds(0);
+          dataSalvandoFim.setHours(horarioSendoSalvo["fim"].split(":", 2)[0])
+          dataSalvandoFim.setMinutes(horarioSendoSalvo["fim"].split(":", 2)[1])
+          dataSalvandoFim.setSeconds(0);
+
+          dataJaCadastradaInicio.setHours(horarioBD["inicio"].split(":", 2)[0])
+          dataJaCadastradaInicio.setMinutes(horarioBD["inicio"].split(":", 2)[1])
+          dataJaCadastradaInicio.setSeconds(0);
+          dataJaCadastradaFim.setHours(horarioBD["fim"].split(":", 2)[0])
+          dataJaCadastradaFim.setMinutes(horarioBD["fim"].split(":", 2)[1])
+          dataJaCadastradaFim.setSeconds(0);
+
+          if ((dataSalvandoInico > dataJaCadastradaInicio) && (dataSalvandoInico < dataJaCadastradaFim)) {
+            this.presentToast('Horário em conflito com outro já cadastrado!');
+            return false;
+          }
+          else if ((dataSalvandoFim > dataJaCadastradaInicio) && (dataSalvandoFim < dataJaCadastradaFim)) {
+            this.presentToast('Horário em conflito com outro já cadastrado!');
+            return false;
+          }
+          else if ((dataSalvandoInico < dataJaCadastradaInicio) && (dataSalvandoFim > dataJaCadastradaFim)) {
+            this.presentToast('Horário em conflito com outro já cadastrado!');
+            return false;
+          }
+          else if ((dataSalvandoInico > dataJaCadastradaInicio) && (dataSalvandoFim < dataJaCadastradaFim)) {
+            this.presentToast('Horário em conflito com outro já cadastrado!');
+            return false;
+          }
+          // Nao pode mesmo dia da semana com horario igual
+          else if ((dataSalvandoInico.toString() === dataJaCadastradaInicio.toString()) || (dataSalvandoFim.toString() === dataJaCadastradaFim.toString())) {
+            this.presentToast('Horário em conflito com outro já cadastrado!');
+            return false;
+          }
+        }
+      }
+    }
+    return true;
   }
 
   // Funcao usada para quando o cliente "Press" na materia
@@ -196,6 +365,36 @@ export class ModalMateriaPage {
         alert("Erro ao criar pasta " + materia + " -- " + JSON.stringify(error));
       }
     }
+  }
+
+  async renomeiaDiretorioMateria(nomeAntigo, nomeNovo) {
+    try {
+      await this.file.moveDir(this.pathNovo, nomeAntigo, this.pathNovo, nomeNovo)
+      //alert("Diretorio "+materia+" ja existe");
+    } catch (error) {
+      alert("Erro ao renomear pasta de " + nomeAntigo + " para " + nomeNovo + " -- Erro: " + JSON.stringify(error));
+    }
+  }
+
+  async isRegistroDuplicado(materiaNome) {
+    try {
+      await this.file.checkDir(this.pathNovo, materiaNome)
+      //alert("Diretorio "+materia+" ja existe");
+      return true;
+    } catch (error) {
+      return false;
+    }
+  }
+
+
+  private presentToast(text) {
+    let toast = this.toastCtrl.create({
+      message: text,
+      duration: 4000,
+      position: 'bottom',
+      cssClass: "toast"
+    });
+    toast.present();
   }
 
 }
